requirements.txt
pyspark>=3.0.0
pandas
scikit-learn
joblib
kafka-python